Update v3
Add role property and change the type of id from Integer to Long

+++ collaborateur/src/main/java/com/Romayssae/collaborateur/repository/TeamRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public interface TeamRepository extends JpaRepository<TeamMember,Integer> {
-        TeamMember findById(int id);
+        TeamMember findById(Long id);
 
 }
